[BOOKINGSG-9152][RYN] add form-label e2e story
